# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "CancelledEvent": {
        "additionalProperties": false,
        "properties": {
          "seq": {
            "minimum": 1.0,
            "title": "Seq",
            "type": "integer"
          },
          "type": {
            "const": "response.cancelled",
            "default": "response.cancelled",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "seq"
        ],
        "title": "CancelledEvent",
        "type": "object"
      },
      "ChatLimits": {
        "additionalProperties": false,
        "properties": {
          "max_message_chars": {
            "maximum": 1048576.0,
            "minimum": 1.0,
            "title": "Max Message Chars",
            "type": "integer"
          },
          "max_response_chars": {
            "maximum": 1048576.0,
            "minimum": 1.0,
            "title": "Max Response Chars",
            "type": "integer"
          },
          "max_stream_bytes": {
            "maximum": 67108864.0,
            "minimum": 1024.0,
            "title": "Max Stream Bytes",
            "type": "integer"
          },
          "stream_idle_timeout_ms": {
            "maximum": 615000.0,
            "minimum": 1000.0,
            "title": "Stream Idle Timeout Ms",
            "type": "integer"
          },
          "stream_timeout_ms": {
            "maximum": 3615000.0,
            "minimum": 1000.0,
            "title": "Stream Timeout Ms",
            "type": "integer"
          }
        },
        "required": [
          "max_message_chars",
          "max_response_chars",
          "max_stream_bytes",
          "stream_timeout_ms",
          "stream_idle_timeout_ms"
        ],
        "title": "ChatLimits",
        "type": "object"
      },
      "ChatRequest": {
        "additionalProperties": false,
        "properties": {
          "attempt": {
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Attempt",
            "type": "integer"
          },
          "conversationId": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_-]+$",
            "title": "Conversationid",
            "type": "string"
          },
          "generationId": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_-]+$",
            "title": "Generationid",
            "type": "string"
          },
          "inputMessageId": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_-]+$",
            "title": "Inputmessageid",
            "type": "string"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/LLMMessage"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Messages",
            "type": "array"
          },
          "modelId": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Modelid",
            "type": "string"
          }
        },
        "required": [
          "generationId",
          "conversationId",
          "inputMessageId",
          "modelId",
          "messages",
          "attempt"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "CompletedEvent": {
        "additionalProperties": false,
        "properties": {
          "seq": {
            "minimum": 1.0,
            "title": "Seq",
            "type": "integer"
          },
          "type": {
            "const": "response.completed",
            "default": "response.completed",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "seq"
        ],
        "title": "CompletedEvent",
        "type": "object"
      },
      "ConversationTitleRequest": {
        "additionalProperties": false,
        "properties": {
          "conversationId": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_-]+$",
            "title": "Conversationid",
            "type": "string"
          },
          "firstMessage": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Firstmessage",
            "type": "string"
          },
          "inputMessageId": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_-]+$",
            "title": "Inputmessageid",
            "type": "string"
          },
          "modelId": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Modelid",
            "type": "string"
          }
        },
        "required": [
          "conversationId",
          "inputMessageId",
          "modelId",
          "firstMessage"
        ],
        "title": "ConversationTitleRequest",
        "type": "object"
      },
      "ConversationTitleResponse": {
        "additionalProperties": false,
        "properties": {
          "conversationId": {
            "title": "Conversationid",
            "type": "string"
          },
          "inputMessageId": {
            "title": "Inputmessageid",
            "type": "string"
          },
          "title": {
            "maxLength": 50,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "conversationId",
          "inputMessageId",
          "title"
        ],
        "title": "ConversationTitleResponse",
        "type": "object"
      },
      "DeltaEvent": {
        "additionalProperties": false,
        "properties": {
          "delta": {
            "title": "Delta",
            "type": "string"
          },
          "seq": {
            "minimum": 1.0,
            "title": "Seq",
            "type": "integer"
          },
          "type": {
            "const": "response.output_text.delta",
            "default": "response.output_text.delta",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "seq",
          "delta"
        ],
        "title": "DeltaEvent",
        "type": "object"
      },
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "FailedEvent": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "seq": {
            "minimum": 1.0,
            "title": "Seq",
            "type": "integer"
          },
          "type": {
            "const": "response.failed",
            "default": "response.failed",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "seq",
          "code",
          "message"
        ],
        "title": "FailedEvent",
        "type": "object"
      },
      "HealthResponse": {
        "additionalProperties": false,
        "properties": {
          "service": {
            "const": "noris-ai",
            "default": "noris-ai",
            "title": "Service",
            "type": "string"
          },
          "status": {
            "const": "ok",
            "default": "ok",
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "LLMMessage": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "maxLength": 1048576,
            "title": "Content",
            "type": "string"
          },
          "role": {
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "content"
        ],
        "title": "LLMMessage",
        "type": "object"
      },
      "LLMModel": {
        "additionalProperties": false,
        "properties": {
          "available": {
            "default": true,
            "title": "Available",
            "type": "boolean"
          },
          "context_window": {
            "default": 8192,
            "maximum": 2000000.0,
            "minimum": 256.0,
            "title": "Context Window",
            "type": "integer"
          },
          "id": {
            "maxLength": 200,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9._/:-]*$",
            "title": "Id",
            "type": "string"
          },
          "max_output_tokens": {
            "default": 1024,
            "maximum": 131072.0,
            "minimum": 1.0,
            "title": "Max Output Tokens",
            "type": "integer"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "provider_limit_evidence": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Limit Evidence"
          },
          "provider_max_output_tokens": {
            "default": 8192,
            "maximum": 131072.0,
            "minimum": 1.0,
            "title": "Provider Max Output Tokens",
            "type": "integer"
          },
          "streaming": {
            "default": true,
            "title": "Streaming",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "LLMModel",
        "type": "object"
      },
      "ModelCatalog": {
        "additionalProperties": false,
        "properties": {
          "default_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Model"
          },
          "limits": {
            "$ref": "#/components/schemas/ChatLimits"
          },
          "models": {
            "items": {
              "$ref": "#/components/schemas/LLMModel"
            },
            "title": "Models",
            "type": "array"
          }
        },
        "required": [
          "models",
          "default_model",
          "limits"
        ],
        "title": "ModelCatalog",
        "type": "object"
      },
      "ReadinessResponse": {
        "additionalProperties": false,
        "properties": {
          "status": {
            "const": "ready",
            "default": "ready",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "ReadinessResponse",
        "type": "object"
      },
      "StartedEvent": {
        "additionalProperties": false,
        "properties": {
          "seq": {
            "minimum": 1.0,
            "title": "Seq",
            "type": "integer"
          },
          "type": {
            "const": "response.started",
            "default": "response.started",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "seq"
        ],
        "title": "StartedEvent",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBasic": {
        "scheme": "basic",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "noris AI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/health/live": {
      "get": {
        "description": "Process health, independent of database availability.",
        "operationId": "getLiveness",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Liveness",
        "tags": [
          "Health"
        ]
      }
    },
    "/api/v1/health/ready": {
      "get": {
        "description": "Ready only when PostgreSQL is reachable and the baseline is applied.",
        "operationId": "getReadiness",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or migrations unavailable"
          }
        },
        "summary": "Readiness",
        "tags": [
          "Health"
        ]
      }
    },
    "/api/v1/llm/chat": {
      "post": {
        "operationId": "streamChat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "discriminator": {
                    "mapping": {
                      "response.cancelled": "#/components/schemas/CancelledEvent",
                      "response.completed": "#/components/schemas/CompletedEvent",
                      "response.failed": "#/components/schemas/FailedEvent",
                      "response.output_text.delta": "#/components/schemas/DeltaEvent",
                      "response.started": "#/components/schemas/StartedEvent"
                    },
                    "propertyName": "type"
                  },
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/StartedEvent"
                    },
                    {
                      "$ref": "#/components/schemas/DeltaEvent"
                    },
                    {
                      "$ref": "#/components/schemas/CompletedEvent"
                    },
                    {
                      "$ref": "#/components/schemas/CancelledEvent"
                    },
                    {
                      "$ref": "#/components/schemas/FailedEvent"
                    }
                  ],
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "408": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Timeout"
          },
          "409": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Content Too Large"
          },
          "415": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "Chat",
        "tags": [
          "LLM"
        ]
      }
    },
    "/api/v1/llm/conversation-title": {
      "post": {
        "operationId": "generateConversationTitle",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationTitleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationTitleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "408": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Timeout"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Content Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "Conversation Title",
        "tags": [
          "LLM"
        ]
      }
    },
    "/api/v1/llm/models": {
      "get": {
        "operationId": "getLLMModels",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelCatalog"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "408": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Timeout"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Content Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "Models",
        "tags": [
          "LLM"
        ]
      }
    }
  }
}